Update collection iteration and RoomIsEmpty calls
Instead of operating on the collection (which pretty much asks for
problems), make a copy of the collection and iterate over that. That
should be MUCH faster and should almost never get modified while
copying, and if it does it should fail much less often.

Every place that I called RoomIsEmpty should have also had a clause for
AllPawnsSleeping... which means it probably should have been factored
out into another function. So I did that and replaced all calls of
either with that function so compatibility with the new night lights
feature is consistent.