Fix bug with free lights
Fix a bug that allowed lights to stay on for free at all times.

Adjust the minimum power draw to be 0.01Wd so that lights (and
especially benches) get the "needs power" icon when power is out. For
benches this prevents pathing to benches where previously a pawn would
repeatedly try to use an unpowered bench.