Update all documentation to use standard docs
Change the old headers to the standard /// documentation to enable
Intellisense for other folks wanting to help out on the mod.

Set the mod up for translation if anyone wants to do that. All UI
display strings from here on out should go through the translation
framework. Console logging will remain in English so that I can
troubleshoot things effectively.