Adjust logging, add mod versioning
Change all informational logging to happen through the new DebugLogger.
This mod should never print to the console unless the debug logging is
enabled (except for the initializing message).

Add a new mod setting that controls whether spammy debug messages get
logged. This will be nice for troubleshooting, but overkill for pretty
much anyone but developers working on this mod.