Revamp Night Lights/Flick Lights Settings
Make it so that the Night Lights settings doesn't take the Flick Lights
setting into account. Previously having Flick Lights off would cause the
pawns to keep the lights on when they sleep, but folks didn't like that
behavior and wanted to be able to use the lights-out-when-sleeping
feature without the normal light flicking feature.