Fix bug with TV handling
Account for the fact that NotifyPatherArrived does not necessarily fire
if a pawn is already at the desired location.

Remove some old documentation headers that didn't need to be there
anymore.

Create a standard way to check if folks are watching a TV and use that
instead (also getting rid of Linq in the process).